agents/1_foundations/community_contributions/chatbot_rag_evaluation/requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
autoflake
backoff
bcrypt
beautifulsoup4
black
blinker
Brotli
build
cachelib
cachetools
certifi
charset-normalizer
chromadb
click
colorama
coloredlogs
contourpy
cycler
dash
dash-bootstrap-components
dash-core-components
dash-design-kit
dash-html-components
dash-mantine-components
dash-table
dash_ag_grid
dataclasses-json
datasets
dill
distro
durationpy
fastapi
ffmpy
filelock
Flask
Flask-Caching
flatbuffers
fonttools
frozenlist
fsspec
gitdb
GitPython
google-auth
google-auth-oauthlib
googleapis-common-protos
gradio
gradio_client
greenlet
gritql
groovy
grpcio
gspread
h11
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
itsdangerous
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kagglehub
kiwisolver
kubernetes
langchain
langchain-chroma
langchain-cli
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langserve
langsmith
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mmh3
mpmath
multidict
multiprocess
mypy-extensions
nest-asyncio
networkx
newsapi-python
newsapi-python-client
nltk
numpy
oauthlib
ollama
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
pathspec
pillow
platformdirs
plotly
posthog
propcache
protobuf
pyarrow
pyasn1
pyasn1_modules
pybase64
pydantic
pydantic-settings
pydantic_core
pydub
pyflakes
pygame
Pygments
pyparsing
PyPDF2
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
retrying
rich
rpds-py
rsa
ruff
safehttpx
safetensors
scikit-learn
scipy
semantic-version
sentence-transformers
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
sse-starlette
starlette
sympy
tenacity
threadpoolctl
tokenizers
tomlkit
torch
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
vizro
watchfiles
websocket-client
websockets
Werkzeug
wrapt
xxhash
yarl
zipp
zstandard
